Add support for Ruby 3.4
GitHub: https://github.com/hardpixel/mrml-ruby/issues/20

Upgrade magnus to 0.6 so that it pulls in the latest rb-sys (0.9.106)
with support for compilation against Ruby 3.4.

Note that this will emit deprecation warnings when using
magnus::memoize!

As magnus 0.6 drops official support for Ruby 2.6 and compilation will
fail with it, we drop it from the list of Ruby versions when cross
compiling.